머지가 머지

+++ src/main/java/org/scoula/mapper/UserMapper.java
@@ -20,7 +20,7 @@ public interface UserMapper {
     void deleteUser(String id);
 
     MemberDTO findByUsername(@Param("username") String username);
-    int countUserByIdx(@Param("usersIdx") Long usersIdx);  //  users_auth 테이블에 해당 유저 정보가 저장되어있는지 확인
+    int countUserByIdx(@Param("usersIdx") int usersIdx);  //  users_auth 테이블에 해당 유저 정보가 저장되어있는지 확인
     void insertUserAuth(@Param("usersIdx") Long usersIdx);
     void insertAuth(AuthDTO authDTO);
     int updatePasswordByUsername(MemberDTO member);
